// types/sound.ts
export enum SoundEvent {
  Boot = "boot",
  Login = "login",
  Shutdown = "shutdown",
  WindowOpen = "window-open",
  WindowClose = "window-close",
  Error = "error",
  Notification = "notification",
  RecycleBin = "recycle-bin",
  Click = "click",
  Hover = "hover",
}

export enum SoundTier {
  Session = "session",
  Dialog = "dialog",
  Interaction = "interaction",
}